requirements.txt
Fabric==1.8.0
IPy==0.81
MySQL-python==1.2.5
PyYAML==3.11
SQLAlchemy==0.8.3
argparse>=1.2.1
boto==2.27.0
dnspython==1.12.0
docopt==0.6.1
ecdsa==0.10
iso8601==0.1.10
netaddr==0.7.12
paramiko==1.12.0
pycrypto==2.6.1
repoze.lru==0.6
requests==2.0.1
simplejson==3.3.1
ssh==1.8.0
wsgiref==0.1.2
